Second part of CMIS-874 (Switch from NSURLConnection to NSURLSession).
The most appropriate task is now used for the request i.e. if a filePath is provided a NSURLSessionDownloadTask is used, uploads are now always done using an NSURLSessionUploadTask and all other requests use NSURLSessionDataTask.

Support for background sessions is going to need some further re-factoring so that will follow in another commit.

git-svn-id: https://svn.apache.org/repos/asf/chemistry/objectivecmis/trunk@1644858 13f79535-47bb-0310-9956-ffa450edef68