Final part of CMIS-874 (Switch from NSURLConnection to NSURLSession).
Several new session parameters have been added to control the use of background sessions. Setting the "kCMISSessionParameterUseBackgroundNetworkSession" key to YES will create a background network session. 

An identifier for the session and the shared container is required when using background mode, these can be set via the "kCMISSessionParameterBackgroundNetworkSessionId" and "kCMISSessionParameterBackgroundNetworkSessionSharedContainerId" keys, respectively. If they are not supplied default values are used, "ObjectiveCMIS" for the background identifier and "ObjectiveCMISContainer" for the shared container.

Two additional session parameter keys have been added to control request timeout (kCMISSessionParameterRequestTimeout) and whether network reachability checks are performed (kCMISSessionParameterCheckNetworkReachability).

git-svn-id: https://svn.apache.org/repos/asf/chemistry/objectivecmis/trunk@1644862 13f79535-47bb-0310-9956-ffa450edef68